src/main/webapp/app/entities/subscribed-clients/delete/subscribed-clients-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, fakeAsync, inject, tick } from '@angular/core/testing';
import { HttpResponse, provideHttpClient } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { SubscribedClientsService } from '../service/subscribed-clients.service';

import { SubscribedClientsDeleteDialogComponent } from './subscribed-clients-delete-dialog.component';

describe('SubscribedClients Management Delete Component', () => {
  let comp: SubscribedClientsDeleteDialogComponent;
  let fixture: ComponentFixture<SubscribedClientsDeleteDialogComponent>;
  let service: SubscribedClientsService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [SubscribedClientsDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(SubscribedClientsDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(SubscribedClientsDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(SubscribedClientsService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
